backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.10.1
pyserial==3.5
pandas==2.1.3
plotly==5.17.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-json-logger==2.0.7